refactoring: move generateDS methods to their own files
