+++ ocrd_models/ocrd_models/ocrd_page_generateds.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 #
-# Generated Thu May 14 15:35:20 2020 by generateDS.py version 2.35.20.
-# Python 3.6.6 (default, Jul 24 2018, 16:39:20)  [GCC 4.9.2]
+# Generated Fri May 15 16:15:27 2020 by generateDS.py version 2.35.20.
+# Python 3.7.6 (default, Jan  8 2020, 19:59:22)  [GCC 7.3.0]
 #
 # Command line options:
 #   ('-f', '')
@@ -16,7 +16,7 @@
 #   repo/assets/data/schema/data/2019.xsd
 #
 # Command line:
-#   /data/monorepo/venv3.6/bin/generateDS -f --root-element="PcGts" -o "ocrd_models/ocrd_models/ocrd_page_generateds.py" --disable-generatedssuper-lookup --user-methods="ocrd_models/ocrd_page_user_methods.py" repo/assets/data/schema/data/2019.xsd
+#   /home/kba/miniconda3/bin/generateDS -f --root-element="PcGts" -o "ocrd_models/ocrd_models/ocrd_page_generateds.py" --disable-generatedssuper-lookup --user-methods="ocrd_models/ocrd_page_user_methods.py" repo/assets/data/schema/data/2019.xsd
 #
 # Current working directory (os.getcwd()):
 #   core
@@ -2850,40 +2850,7 @@ class PageType(GeneratedsSuper):
             obj_.original_tagname_ = 'CustomRegion'
     def __hash__(self):
         return hash(self.id)
-
-    def get_AllRegions(self, regions=None, reading_order=False):
-        """"
-        Get all the *Region element or only those provided by ``regions``.
-        Returned in random order unless ``reading_order`` is set (NOT CURRENTLY IMPLEMENTED)
-        """
-        if reading_order:
-            reading_order = self.get_ReadingOrder()
-        if not regions:
-            regions = ['Advert', 'Chart', 'Chem', 'Custom', 'Graphic', 'Image', 'LineDrawing', 'Map', 'Maths', 'Music', 'Noise', 'Separator', 'Table', 'Text', 'Unknown']
-        ret = []
-        for region in regions:
-            ret += getattr(self, 'get_{}Region'.format(region))()
-        if reading_order:
-            reading_order = reading_order.get_OrderedGroup() or reading_order.get_UnorderedGroup()
-        if reading_order:
-            def get_recursive_reading_order(rogroup):
-                if isinstance(rogroup, (OrderedGroupType, OrderedGroupIndexedType)):
-                    elements = rogroup.get_AllIndexed()
-                if isinstance(rogroup, (UnorderedGroupType, UnorderedGroupIndexedType)):
-                    elements = (rogroup.get_RegionRef() + rogroup.get_OrderedGroup() + rogroup.get_UnorderedGroup())
-                regionrefs = list()
-                for elem in elements:
-                    regionrefs.append(elem.get_regionRef())
-                    if not isinstance(elem, (RegionRefType, RegionRefIndexedType)):
-                        regionrefs.extend(get_recursive_reading_order(elem))
-                return regionrefs
-            reading_order = get_recursive_reading_order(reading_order)
-        if reading_order:
-            ret = dict([(region.id, region) for region in ret])
-            return [ret[region_id] for region_id in reading_order if region_id in ret]
-        else:
-            return ret
-    # end class PageType
+# end class PageType
 
 
 class CoordsType(GeneratedsSuper):
@@ -5380,7 +5347,28 @@ class OrderedGroupIndexedType(GeneratedsSuper):
             obj_.original_tagname_ = 'UnorderedGroupIndexed'
     def __hash__(self):
         return hash(self.id)
-
+    def get_AllIndexed(self):
+        return sorted(self.get_RegionRefIndexed() + self.get_OrderedGroupIndexed() + self.get_UnorderedGroupIndexed(), key=lambda x : x.index)
+    
+    def clear_AllIndexed(self):
+        ret = self.get_AllIndexed()
+        self.set_RegionRefIndexed([])
+        self.set_OrderedGroupIndexed([])
+        self.set_UnorderedGroupIndexed([])
+        return ret
+    
+    def add_AllIndexed(self, elements):
+        if not isinstance(elements, list):
+            elements = [elements]
+        for element in sorted(elements, key=lambda x : x.index):
+            if isinstance(element, RegionRefIndexedType):
+                self.add_RegionRefIndexed(element)
+            elif isinstance(element, OrderedGroupIndexedType):
+                self.add_OrderedGroupIndexed(element)
+            elif isinstance(element, UnorderedGroupIndexedType):
+                self.add_UnorderedGroupIndexed(element)
+        return self.get_AllIndexed()
+    
     def exportChildren(self, outfile, level, namespaceprefix_='', namespacedef_='xmlns:pc="http://schema.primaresearch.org/PAGE/gts/pagecontent/2019-07-15"', name_='OrderedGroupType', fromsubclass_=False, pretty_print=True):
         eol_ = '\n' if pretty_print else ''
         namespaceprefix_ = 'pc:'
@@ -5400,27 +5388,7 @@ class OrderedGroupIndexedType(GeneratedsSuper):
                 cleaned.append(entry)
         for entry in cleaned:
             entry.export(outfile, level, namespaceprefix_, namespacedef_='', name_=entry.__class__.__name__[:-4], pretty_print=pretty_print)
-
-    def get_AllIndexed(self):
-        return sorted(self.get_RegionRefIndexed() + self.get_OrderedGroupIndexed() + self.get_UnorderedGroupIndexed(), key=lambda x : x.index) 
-    def add_AllIndexed(self, elements):
-        if not isinstance(elements, list):
-            elements = [elements]
-        for element in sorted(elements, key=lambda x : x.index):
-            if isinstance(element, RegionRefIndexedType):
-                self.add_RegionRefIndexed(element)
-            elif isinstance(element, OrderedGroupIndexedType):
-                self.add_OrderedGroupIndexed(element)
-            elif isinstance(element, UnorderedGroupIndexedType):
-                self.add_UnorderedGroupIndexed(element)
-        return self.get_AllIndexed()
-
-    def clear_AllIndexed(self):
-        ret = self.get_AllIndexed()
-        self.set_RegionRefIndexed([])
-        self.set_OrderedGroupIndexed([])
-        self.set_UnorderedGroupIndexed([])
-        return ret
+    
 # end class OrderedGroupIndexedType
 
 
@@ -6169,7 +6137,28 @@ class OrderedGroupType(GeneratedsSuper):
             obj_.original_tagname_ = 'UnorderedGroupIndexed'
     def __hash__(self):
         return hash(self.id)
-
+    def get_AllIndexed(self):
+        return sorted(self.get_RegionRefIndexed() + self.get_OrderedGroupIndexed() + self.get_UnorderedGroupIndexed(), key=lambda x : x.index)
+    
+    def clear_AllIndexed(self):
+        ret = self.get_AllIndexed()
+        self.set_RegionRefIndexed([])
+        self.set_OrderedGroupIndexed([])
+        self.set_UnorderedGroupIndexed([])
+        return ret
+    
+    def add_AllIndexed(self, elements):
+        if not isinstance(elements, list):
+            elements = [elements]
+        for element in sorted(elements, key=lambda x : x.index):
+            if isinstance(element, RegionRefIndexedType):
+                self.add_RegionRefIndexed(element)
+            elif isinstance(element, OrderedGroupIndexedType):
+                self.add_OrderedGroupIndexed(element)
+            elif isinstance(element, UnorderedGroupIndexedType):
+                self.add_UnorderedGroupIndexed(element)
+        return self.get_AllIndexed()
+    
     def exportChildren(self, outfile, level, namespaceprefix_='', namespacedef_='xmlns:pc="http://schema.primaresearch.org/PAGE/gts/pagecontent/2019-07-15"', name_='OrderedGroupType', fromsubclass_=False, pretty_print=True):
         eol_ = '\n' if pretty_print else ''
         namespaceprefix_ = 'pc:'
@@ -6189,27 +6178,7 @@ class OrderedGroupType(GeneratedsSuper):
                 cleaned.append(entry)
         for entry in cleaned:
             entry.export(outfile, level, namespaceprefix_, namespacedef_='', name_=entry.__class__.__name__[:-4], pretty_print=pretty_print)
-
-    def get_AllIndexed(self):
-        return sorted(self.get_RegionRefIndexed() + self.get_OrderedGroupIndexed() + self.get_UnorderedGroupIndexed(), key=lambda x : x.index) 
-    def add_AllIndexed(self, elements):
-        if not isinstance(elements, list):
-            elements = [elements]
-        for element in sorted(elements, key=lambda x : x.index):
-            if isinstance(element, RegionRefIndexedType):
-                self.add_RegionRefIndexed(element)
-            elif isinstance(element, OrderedGroupIndexedType):
-                self.add_OrderedGroupIndexed(element)
-            elif isinstance(element, UnorderedGroupIndexedType):
-                self.add_UnorderedGroupIndexed(element)
-        return self.get_AllIndexed()
-
-    def clear_AllIndexed(self):
-        ret = self.get_AllIndexed()
-        self.set_RegionRefIndexed([])
-        self.set_OrderedGroupIndexed([])
-        self.set_UnorderedGroupIndexed([])
-        return ret
+    
 # end class OrderedGroupType
