+++ ocrd_models/ocrd_page_user_methods/__hash__.py
@@ -1,0 +1,2 @@
+def __hash__(self):
+    return hash(self.id)
