+++ ocrd_models/ocrd_page_user_methods/clear_AllIndexed.py
@@ -1,0 +1,7 @@
+def clear_AllIndexed(self):
+    ret = self.get_AllIndexed()
+    self.set_RegionRefIndexed([])
+    self.set_OrderedGroupIndexed([])
+    self.set_UnorderedGroupIndexed([])
+    return ret
+
