+++ ocrd_models/ocrd_page_user_methods/get_AllIndexed.py
@@ -1,0 +1,3 @@
+def get_AllIndexed(self):
+    return sorted(self.get_RegionRefIndexed() + self.get_OrderedGroupIndexed() + self.get_UnorderedGroupIndexed(), key=lambda x : x.index)
+
